pyproject.toml
[project]
name = "avatar-train"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.20.3",
    "boto3>=1.35.85",
    "deepface>=0.0.93",
    "diffusers>=0.31.0",
    "face-detection>=0.2.2",
    "facenet-pytorch>=2.5.3",
    "ffmpeg-python>=0.2.0",
    "insightface>=0.7.3",
    "llvmlite>=0.43.0",
    "loguru>=0.7.2",
    "lpips>=0.1.4",
    "matplotlib>=3.9.3",
    "mediapipe>=0.10.18",
    "onnxruntime>=1.20.1",
    "opencv-python-headless>=4.10.0.84",
    "p-tqdm>=1.4.2",
    "pydantic>=2.9.2",
    "pymp>=0.0.6",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "soundfile>=0.12.1",
    "tf-keras>=2.18.0",
    "torch-fidelity>=0.3.0",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchmetrics>=1.6.0",
    "torchvision>=0.20.1",
    "transformers>=4.31.0",
]

[project.optional-dependencies]
wandb = [
    "wandb>=0.18.7",
]

[dependency-groups]
loss = [
    "lpips>=0.1.4",
]
dev = [
    "clearml>=1.16.5",
    "deepface>=0.0.93",
    "ruff>=0.7.4",
    "tensorflow>=2.18.0",
]
